Allow multiple buttons at the bottom per planet
This ends the requirement that there be one button at the bottom for each
available planet. Instead, the buttons are created dynamically from the POI
data we load. This restructures the POI data, so each group of points specifies
which planet it belongs to.

There's still a problem whereby the planet icon shrinks when it's highlighted.
That's for another day.