Resurrect planet system
This change reintroduces the planet view and controller as the means for
switching between batches of POIs. This is consistent with the design of the
original touchscreen, but had been removed for this version in favor of a less
easily customizable version that, notably, didn't interact well with some
changes I just finished on the attract loop. Also introduces a "whichpage" view
(yeah, it's a stupid name; feel free to change it if you care that much) to
switch between map mode and "diving" mode, which is what the original planet
controller had sort of been half adulterated to handle.

Change-Id: I3f4c492615c75014a5e7fb99c47caaa3bf0bcf9e